:recycle: refact(common) : 예외처리 관련 로직 개편

+++ src/main/java/org/quack/QUACKServer/review/controller/ReviewController.java
@@ -3,14 +3,16 @@ package org.quack.QUACKServer.review.controller;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.quack.QUACKServer.auth.domain.CustomerUserInfo;
+import org.quack.QUACKServer.core.error.constant.ErrorCode;
+import org.quack.QUACKServer.core.error.exception.CommonException;
 import org.quack.QUACKServer.review.dto.request.CreateReviewRequest;
 import org.quack.QUACKServer.review.dto.response.MyReviewResponse;
 import org.quack.QUACKServer.review.dto.response.ReviewInitResponse;
 import org.quack.QUACKServer.review.enums.ReviewType;
 import org.quack.QUACKServer.review.service.ReviewService;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
-import static org.quack.QUACKServer.auth.domain.PrincipalManager.getAuthenticatedUser;
 
 @RestController
 @Slf4j
@@ -26,25 +28,34 @@ public class ReviewController {
     }
 
     @PostMapping("/{restaurantId}/create")
-    public String createReview(@PathVariable Long restaurantId,
-                               @RequestBody CreateReviewRequest request) {
-        CustomerUserInfo loginUser = getAuthenticatedUser();
+    public String createReview(
+            @AuthenticationPrincipal CustomerUserInfo loginUser,
+            @PathVariable Long restaurantId, @RequestBody CreateReviewRequest request) {
 
+        if(loginUser == null) {
+            throw new CommonException(ErrorCode.UNAUTHORIZED_USER);
+        }
         return reviewService.createReview(loginUser, restaurantId, request);
     }
 
     @PostMapping("/{reviewId}/delete")
-    public String deleteReview(@PathVariable Long reviewId) {
-        CustomerUserInfo loginUser = getAuthenticatedUser();
-
+    public String deleteReview(
+            @AuthenticationPrincipal CustomerUserInfo loginUser,
+            @PathVariable Long reviewId) {
+        if(loginUser == null) {
+            throw new CommonException(ErrorCode.UNAUTHORIZED_USER);
+        }
         return reviewService.deleteReview(loginUser, reviewId);
     }
 
     @GetMapping("/my-reviews")
-    public MyReviewResponse getMyReviews(@RequestParam(defaultValue = "0") int pageNum,
-                                         @RequestParam(defaultValue = "10") int sizeNum){
-        CustomerUserInfo loginUser = getAuthenticatedUser();
-
+    public MyReviewResponse getMyReviews(
+            @AuthenticationPrincipal CustomerUserInfo loginUser,
+            @RequestParam(defaultValue = "0") int pageNum,
+            @RequestParam(defaultValue = "10") int sizeNum){
+        if(loginUser == null) {
+            throw new CommonException(ErrorCode.UNAUTHORIZED_USER);
+        }
         return reviewService.getMyReviews(loginUser, pageNum, sizeNum);
     }
 }
